requirements-dev.txt
numpy==1.24.4
pandas==2.3.1
seaborn==0.13.2
matplotlib==3.10.3
scipy==1.16.1
scikit-learn==1.7.0
tensorflow==2.17.1
tensorflow-intel==2.17.1
nltk==3.9.1
PyYAML==6.0.2
pymongo==4.14.0
pillow==11.3.0
python-box==7.3.2
python-dotenv==1.1.1
ensure==1.0.4
evidently==0.7.12
lime==0.2.0.1
prometheus_client==0.22.1
xgboost==3.0.2
uvicorn==0.35.0
gunicorn==23.0.0
dvc==3.61.0
mlflow==2.22.1
dagshub==0.5.10
fastapi==0.116.1
pydantic==2.11.7
tqdm==4.67.1
requests==2.32.4
kaggle==1.7.4.5
kagglehub==0.3.12
jupyter_core==5.8.1
Jinja2==3.1.6
ipykernel==6.29.5
ipython==9.4.0
pytest==8.4.1
flake8==7.3.0
boto3==1.39.4
black==25.1.0
tox==4.11.3
mypy==1.17.0